Add files via upload
Sequence Classification with LSTM ,RNN in Python with Keras
In this project, we are going to work on Sequence to Sequence Prediction using IMDB Movie Review Data set​ using Keras in Python
A sequence to sequence prediction for developing a classification system is of very much required in developing applications. Standard approaches for developing applications won't help in providing accuracy. Hence, as an example let's take an IMDB movie review data set and create some benchmarks by using RNN, RNN with LSTM and drop out rate, RNN with CNN, and RNN with CNN plus drop out rate to make a composite sequence to sequence classification work. We can compare the model accuracy as well.